[MINOR][PYTHON][TESTS] Consolidate DataStreamReader.name() tests into test_streaming.py
